Move iterators into their own header file
